backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
httptools==0.7.1
idna==3.11
matplotlib==3.7.5
numpy==2.3.3
openpyxl==3.1.5
orjson==3.11.7
pandas==2.3.3
pyarrow==19.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.8.2
python-multipart==0.0.22
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1; platform_system != "Windows"
watchfiles==1.11.1

# Test dependencies
httpx>=0.27
pytest>=8.0
